+++ apps/web/src/views/BuyCrypto/components/AccordionDropdown/AccordionItem.tsx
@@ -60,22 +60,22 @@ function AccordionItem({
     currentLanguage: { locale },
   } = useTranslation()
   const contentRef = useRef<HTMLDivElement>(null)
-  const [height, setHeight] = useState(active ? 240 : 96)
+  const [maxHeight, setMaxHeight] = useState(active ? 500 : 150)
   const multiple = false
-  const [visiblity, setVisiblity] = useState(false)
+  const [visibility, setVisibility] = useState(false)
   const [mobileTooltipShow, setMobileTooltipShow] = useState(false)
   const currentTimestamp = Math.floor(Date.now() / 1000)
   const { days, hours, minutes, seconds } = getTimePeriods(currentTimestamp - CURRENT_CAMPAIGN_TIMESTAMP)
-  const isActive = () => (multiple ? visiblity : active)
+  const isActive = () => (multiple ? visibility : active)
 
-  const toogleVisiblity = useCallback(() => {
-    setVisiblity((v) => !v)
+  const toggleVisibility = useCallback(() => {
+    setVisibility((v) => !v)
     btnOnClick()
-  }, [setVisiblity, btnOnClick])
+  }, [setVisibility, btnOnClick])
 
   useEffect(() => {
     const contentEl = getRefValue(contentRef)
-    setHeight(contentEl?.scrollHeight + 96)
+    setMaxHeight(contentEl?.scrollHeight + 150)
   }, [active])
 
   const {
@@ -100,8 +100,8 @@ function AccordionItem({
   if (quote.amount === 0) {
     return (
       <Flex flexDirection="column">
-        <CryptoCard padding="16px 16px" style={{ height: '70px' }} position="relative" isClicked={false} isDisabled>
-          <RowBetween paddingBottom="24px" paddingTop="8px">
+        <CryptoCard position="relative" isClicked={false} isDisabled>
+          <RowBetween>
             <OnRampProviderLogo provider={quote.provider} />
             <TooltipText
               ref={buyCryptoTargetRef}
@@ -126,15 +126,14 @@ function AccordionItem({
     <Flex flexDirection="column">
       <CryptoCard
         padding="18px 18px"
-        style={{ height }}
-        onClick={!isActive() ? toogleVisiblity : () => null}
+        style={{ maxHeight }}
+        onClick={!isActive() ? toggleVisibility : () => null}
         position="relative"
         isClicked={active}
         isDisabled={false}
       >
         <RowBetween>
           <OnRampProviderLogo provider={quote.provider} />
-
           <Text ml="4px" fontSize="18px" color="#7A6EAA" fontWeight="bold">
             {formatLocaleNumber({
               number: quote.quote,