+++ apps/web/src/views/Nft/market/Collection/IndividualNFTPage/shared/ManageNFTsCard.tsx
@@ -18,7 +18,7 @@ import { NftLocation, NftToken, Collection } from 'state/nftMarket/types'
 import { formatNumber } from '@pancakeswap/utils/formatBalance'
 import ConnectWalletButton from 'components/ConnectWalletButton'
 import { useTranslation } from '@pancakeswap/localization'
-import { isAddress } from 'utils'
+import { safeGetAddress } from 'utils'
 import { CollectibleRowContainer, SmallRoundedImage } from './styles'
 import ProfileNftModal from '../../../components/ProfileNftModal'
 import SellModal from '../../../components/BuySellModals/SellModal'
@@ -149,7 +149,7 @@ interface ManageNftsCardProps {
 const getNftFilter = (location: NftLocation) => {
   return (nft: NftToken, collectionAddress: string, tokenId: string | number): boolean => {
     return (
-      isAddress(nft.collectionAddress) === isAddress(collectionAddress) &&
+      safeGetAddress(nft.collectionAddress) === safeGetAddress(collectionAddress) &&
       (tokenId ? nft.attributes[0].value === tokenId : true) &&
       nft.location === location
     )