+++ apps/web/src/views/V3Info/components/Search/index.tsx
@@ -12,7 +12,7 @@ import { styled } from 'styled-components'
 import { formatAmount } from 'utils/formatInfoNumbers'
 import { CurrencyLogo, DoubleCurrencyLogo } from 'views/Info/components/CurrencyLogo'
 
-import { isAddress } from 'utils'
+import { safeGetAddress } from 'utils'
 import { v3InfoPath } from '../../constants'
 import { usePoolsData, useSearchData, useTokensData } from '../../hooks'
 import { PoolData } from '../../types'
@@ -324,10 +324,8 @@ const Search = () => {
                     <Flex>
                       <CurrencyLogo address={token.address} chainName={chainName} />
                       <Text ml="10px">
-                        <Text>{`${
-                          (token.address && subgraphTokenName[isAddress(token.address) || undefined]) || token.name
-                        } (${
-                          (token.address && subgraphTokenSymbol[isAddress(token.address) || undefined]) || token.symbol
+                        <Text>{`${(token.address && subgraphTokenName[safeGetAddress(token.address)]) || token.name} (${
+                          (token.address && subgraphTokenSymbol[safeGetAddress(token.address)]) || token.symbol
                         })`}</Text>
                       </Text>
                       {/* <SaveIcon
@@ -395,8 +393,8 @@ const Search = () => {
                         chainName={chainName}
                       />
                       <Text ml="10px" style={{ whiteSpace: 'nowrap' }}>
-                        <Text>{`${subgraphTokenSymbol[isAddress(p.token0.address) || undefined] ?? p.token0.symbol} / ${
-                          subgraphTokenSymbol[isAddress(p.token1.address) || undefined] ?? p.token1.symbol
+                        <Text>{`${subgraphTokenSymbol[safeGetAddress(p.token0.address)] ?? p.token0.symbol} / ${
+                          subgraphTokenSymbol[safeGetAddress(p.token1.address)] ?? p.token1.symbol
                         }`}</Text>
                       </Text>
                       <GreyBadge ml="10px" style={{ fontSize: 14 }}>