+++ packages/smart-router/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @pancakeswap/smart-router
 
+## 4.9.1
+
+### Patch Changes
+
+- Updated dependencies [8e3ac5427]
+  - @pancakeswap/sdk@5.7.0
+  - @pancakeswap/multicall@3.2.1
+  - @pancakeswap/tokens@0.5.1
+  - @pancakeswap/v3-sdk@3.5.1
+
 ## 4.9.0
 
 ### Minor Changes